:nail_care: spring clean all the badges
Sorry @bokub, I was super hyped when you introduced badges, but now I
feel like it is deteriorating the readability of the content (the
links).
I tried to align the links on the right, but it would mean adding loads
of html, and impact the readability of the source.
So for now, no more badges :disappointed: